feat: [부서]수정, 구성원 검색 및 수정 기능 구현
- 부서 수정 기능 구현
- 부서 매니저 검색 기능 구현
- 부서 구성원 검색 및 수정 기능 구현
- ExceptionMsg.java enum 생성
- DeptDto에 CDR_DEPT_CNT 속성 추가
- 부서 입력 및 수정 시 파라미터 mngr_empl_no에서 empl_id로 변경
- 부서 삭제 시 하위부서 여부 확인 기능 추가
- 부서 조회 시 하위부서 수 조회 기능 구현

+++ src/main/java/site/roombook/service/DeptServiceImpl.java
@@ -1,26 +1,34 @@
 package site.roombook.service;
 
+import org.apache.logging.log4j.LogManager;
+import org.apache.logging.log4j.Logger;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.dao.DuplicateKeyException;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import site.roombook.ExceptionMsg;
+import site.roombook.dao.BlngDeptDao;
 import site.roombook.dao.DeptDao;
 import site.roombook.dao.EmplDao;
-import site.roombook.domain.DeptDto;
-import site.roombook.domain.EmplDto;
+import site.roombook.domain.*;
 
+import java.lang.reflect.InvocationTargetException;
 import java.util.*;
 import java.util.concurrent.ThreadLocalRandom;
 
 @Service
 public class DeptServiceImpl implements DeptService {
-
+    private static final Logger logger = LogManager.getLogger();
 
     @Autowired
     DeptDao deptDao;
 
     @Autowired
     EmplDao emplDao;
+
+    @Autowired
+    BlngDeptDao blngDeptDao;
 
     @Override
     public boolean haveIdenticalDeptNm(String deptNm) {
@@ -29,12 +37,12 @@ public class DeptServiceImpl implements DeptService {
 
     @Transactional
     @Override
-    public boolean saveOneDept(List<DeptDto> list, String emplNo) throws DuplicateKeyException, NullPointerException{
+    public boolean saveOneDept(List<DeptDto> depts, String emplNo) throws DuplicateKeyException, NullPointerException{
         // TODO : test하기, service test 돌아가게 수정, test갯수는 필요한 것만 최소화, 중복 제거하기
         DeptDto newDept = null;
         int rowCnt;
 
-        for (DeptDto deptDto : list) {
+        for (DeptDto deptDto : depts) {
             if(Objects.equals(deptDto.getDEPT_CD(), NO_DEPT_CD)){
                 newDept = deptDto;
                 break;
@@ -51,10 +59,10 @@ public class DeptServiceImpl implements DeptService {
             return false;
         }
 
-        list.forEach( deptDto -> deptDto.setLAST_UPDR_IDNF_NO(emplNo));
+        depts.forEach( deptDto -> deptDto.setLAST_UPDR_IDNF_NO(emplNo));
 
         try{
-            deptDao.updateAllDeptTreeOdrData(list);
+            deptDao.updateAllDeptTreeOdrData(depts);
         } catch (Exception e){
             e.printStackTrace();
         }
@@ -101,11 +109,15 @@ public class DeptServiceImpl implements DeptService {
 
     @Transactional
     @Override
-    public void deleteDeptWithNoEmpl(String deptCd) throws RuntimeException{
+    public void deleteDept(String deptCd) throws IllegalArgumentException{
+        if(deptDao.selectCdrDeptCnt(deptCd)!=0){
+            throw new IllegalArgumentException(ExceptionMsg.DEPT_DELETE_SUB_DEPT.getCode());
+        }
+
         int deletedRowCnt = deptDao.deleteDeptWithNoEmpl(deptCd);
 
         if(deletedRowCnt==0){
-            throw new RuntimeException("Dept with members cannot be deleted");
+            throw new IllegalArgumentException(ExceptionMsg.DEPT_DELETE_MEMBER.getCode());
         }
 
     }
@@ -128,5 +140,70 @@ public class DeptServiceImpl implements DeptService {
     @Override
     public EmplDto getDeptMngr(String emplNo){
         return emplDao.selectOneEmplProfile(emplNo);
+    }
+
+    @Transactional
+    @Override
+    public List<EmplDto> searchEmplWithRnmOrEmail(String keyword){
+        return emplDao.selectEmplProfilesWithRnmOrEmail(keyword);
+    }
+
+    @Transactional
+    @Override
+    public boolean modifyOneDept(Map<String, String> deptDataAndEmplId){
+        return deptDao.updateDept(deptDataAndEmplId)>0;
+    }
+
+    @Override
+    public List<DeptAndEmplDto> getProfilesOfMemberAndDeptName(String deptCd){
+        return deptDao.selectMemberProfilesAndDeptName(deptCd);
+    }
+
+    @Override
+    @Transactional
+    public void modifyDeptMem(String deptCd, List<String> memIDs, String modifier) throws DataIntegrityViolationException {
+        List<DeptAndEmplDto> oldMemList = deptDao.selectMemberProfilesAndDeptName(deptCd);
+
+        HashSet<String> newMemIDs = new HashSet<>(memIDs);
+        HashSet<String> oldMemIDs = new HashSet<>();
+        for (DeptAndEmplDto de : oldMemList) {
+            oldMemIDs.add(de.getEMPL_ID());
+        }
+
+        HashSet<String> memForRemoval = new HashSet<>(oldMemIDs);
+        memForRemoval.removeAll(newMemIDs);
+
+        if(!memForRemoval.isEmpty()){
+            Map<String, Object> blngDepts = new HashMap<>();
+            blngDepts.put("BLNG_DEPT_CD", deptCd);
+            blngDepts.put("emplIDs", memForRemoval.toArray());
+
+            try{
+                blngDeptDao.deleteBlngDepts(blngDepts);
+            } catch (InvocationTargetException e){
+                logger.info("dept 구성원 수정 중 delete value 없음");
+            }
+        }
+
+        newMemIDs.removeAll(oldMemIDs);
+
+        if(!newMemIDs.isEmpty()){
+            List<BlngDeptAndEmplIdDto> newMemList = new ArrayList<>();
+
+            for (String newMemID : newMemIDs) {
+                newMemList.add(new BlngDeptAndEmplIdDto(deptCd, newMemID, modifier, modifier));
+            }
+
+            try{
+                blngDeptDao.insertBlngDepts(newMemList);
+            } catch (InvocationTargetException e){
+                logger.info("dept 구성원 수정 중 insert value 없음");
+            }
+        }
+    }
+
+    @Override
+    public DeptAndEmplDto getDeptDetailInfo(String deptCd){
+        return deptDao.selectOneDeptAndMngrAndCdrDeptCnt(deptCd);
     }
 }
